refactoring

+++ src/main/java/com/kine/Kine/food/service/service/FoodService.java
@@ -1,5 +1,7 @@
 package com.kine.Kine.food.service.service;
+import com.kine.Kine.food.service.dto.GetFoodDTO;
 import com.kine.Kine.food.service.dto.UpdateDTO;
+
 import com.kine.Kine.food.service.repository.FoodRepository;
 import com.kine.Kine.food.service.repository.Food;
 import jakarta.persistence.Id;
@@ -10,6 +12,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.BeanUtils;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
+
 @Service
 public class FoodService {
     private final FoodRepository foodRepository;
@@ -17,27 +22,30 @@ public class FoodService {
     public FoodService(FoodRepository foodRepository){
         this.foodRepository = foodRepository;
     }
-    public Iterable<FoodDTO> getFood() {
+    public Iterable<GetFoodDTO> getFood() {
         Iterable<Food> foundFood = foodRepository.findAll();
-        List<FoodDTO> foodList = new ArrayList<>();
+        List<GetFoodDTO> foodList = new ArrayList<>();
         for (Food food : foundFood)
-            foodList.add(new FoodDTO(food.getId(), food.getName(), food.getCalories(), food.getProteins(), food.getCarbs(), food.getFat()));
+            foodList.add( food.ConvertToGETFoodDTO() );
         return foodList;
     }
 
-    public Iterable<FoodDTO> findFoodByName(String name){
+    public Iterable<GetFoodDTO> findFoodByName(String name){
         Iterable<Food> foundFood = foodRepository.findFoodByName(name);
-        List<FoodDTO> foodList = new ArrayList<>();
+        List<GetFoodDTO> foodList = new ArrayList<>();
         for (Food food : foundFood)
-            foodList.add(new FoodDTO(food.getId(), food.getName(), food.getCalories(), food.getProteins(), food.getCarbs(), food.getFat()));
+            foodList.add(food.ConvertToGETFoodDTO());
         return foodList;
     }
-    public Iterable<FoodDTO> findFoodById(long id){
-        Food foundFood = foodRepository.findFoodById(id);
-        List<FoodDTO> foodList = new ArrayList<>();
-
-        foodList.add(new FoodDTO(foundFood.getId(), foundFood.getName(), foundFood.getCalories(), foundFood.getProteins(), foundFood.getCarbs(), foundFood.getFat()));
-        return foodList;
+    public Iterable<GetFoodDTO> findFoodById(long id){
+        try {
+            Food foundFood = foodRepository.findFoodById(id);
+            List<GetFoodDTO> foodList = new ArrayList<>();
+            foodList.add(foundFood.ConvertToGETFoodDTO());
+            return foodList;
+        }catch (Exception e){
+            return new ArrayList<>();
+        }
     }
     public FoodDTO createFood(FoodDTO foodDTO){
         Food food = foodRepository.save(new Food(foodDTO.getName(), foodDTO.getCalories(), foodDTO.getProteins(), foodDTO.getCarbs(), foodDTO.getFat()));
@@ -69,7 +77,12 @@ public class FoodService {
         return new FoodDTO(food.getId(), food.getName(), food.getCalories(), food.getProteins(), food.getCarbs(), food.getFat());
     }
     public void deleteFoodById(Long id) {
-        foodRepository.deleteById(id);
+        Optional<Food> foodOptional = foodRepository.findById(id);
+        if (foodOptional.isPresent()) {
+            foodRepository.deleteById(id);
+        } else {
+            throw new NoSuchElementException("Food item with id " + id + " not found");
+        }
     }
